Move notify functions: basics.js -> notify.js
